pyproject.toml
[project]
name = "dqx sample profiling"
version = "0.1.9"
requires-python = ">=3.10"
dependencies = []

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "databricks-labs-dqx>=0.13.0"
]